refactor: extract menu preference helpers
